APP Structure/Backend/requirements.txt
# Flask and Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
python-socketio==5.10.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Authentication
PyJWT==2.8.0
bcrypt==4.1.1

# Redis (Caching)
redis==5.0.1
Flask-Caching==2.1.0

# Machine Learning
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
tensorflow==2.15.0
xgboost==2.0.3
joblib==1.3.2

# Data Processing
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0

# Blockchain
web3==6.11.3
eth-account==0.10.0
eth-utils==2.3.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-mock==3.12.0

# Code Quality
black==23.12.1
flake8==6.1.0
pylint==3.0.3
mypy==1.7.1

# Monitoring and Logging
python-json-logger==2.0.7
sentry-sdk==1.39.1

# Task Queue (Optional)
celery==5.3.4
kombu==5.3.4

# Production Server
gunicorn==21.2.0
gevent==23.9.1
